Refactor presenter to be concerned only with UI updates
Create a new class that will take care of finding the current location, and task the presenter with calling a new class in the view to show a progress dialog while the location is being found. Anticipate that the presenter will have multiple similar responsibilities controlling the view and want to separate concerns.