Run each row of daily forecast through a view model
Don't want to use a lot of static binding methods because another view model is already set up, and can reuse some of the methods written for it. Methods remain testable if they aren't turned into binding adapters. Downside is the creation of a set of 32 objects every time the recycler view updates